# LibGfx/ICC: Add a one-element cache for CLUT conversions